# web/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Referral API",
        "description": "This is API Microservice Referral Program",
        "contact": {
            "name": "Support Team",
            "email": "admin@sumra.net"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/referrals/admin/referrals-list": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Method for get list all referral users",
                "description": "Get referral users",
                "operationId": "App\\Api\\V1\\Controllers\\Admin\\UsersController::index",
                "parameters": [
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Order By",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderDirection",
                        "in": "query",
                        "description": "Order Direction",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search keywords",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bulk",
                        "in": "query",
                        "description": "Bulk filter",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "default": [
                            "ManagerRead",
                            "User",
                            "ManagerWrite"
                        ]
                    }
                ],
                "x-auth-type": "Application & Application User",
                "x-throttling-tier": "Unlimited",
                "x-wso2-application-security": {
                    "security-types": [
                        "oauth2"
                    ],
                    "optional": "false"
                }
            }
        },
        "/v1/referrals/admin/referrals-list/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get detail info about user",
                "description": "Get detail info about user",
                "operationId": "App\\Api\\V1\\Controllers\\Admin\\UsersController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data of user"
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Get detail info about user",
                                            "properties": {
                                                "code": {
                                                    "description": "code of error",
                                                    "type": "string"
                                                },
                                                "message": {
                                                    "description": "error message",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "default": [
                            "ManagerRead",
                            "User",
                            "ManagerWrite"
                        ]
                    }
                ],
                "x-auth-type": "Application & Application User",
                "x-throttling-tier": "Unlimited",
                "x-wso2-application-security": {
                    "security-types": [
                        "oauth2"
                    ],
                    "optional": "false"
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "default": {
                "type": "oauth2",
                "description": "Auth Scheme",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "https://is.sumra.net/oauth2",
                        "scopes": {
                            "ManagerRead": "Manager can read",
                            "User": "User access",
                            "ManagerWrite": "Manager can write"
                        }
                    }
                }
            }
        }
    }
}